use condition build for os vendor syscall
